Spec links for assay metadata corrected
These "Spec" links incorrectly pointed to the assay docs in Github which contain the links to the corresponding "metadata specs" Github pages which contain the links to download the corresponding tsvs. 

I changed the "Spec" links so that they now point directly to the "metadata specs" Github page with the specifications and the tsv links.