Update README.md
Adding identifiers of folks responsible for reviewing issues and managing questions relating to specific documentation content areas.  We can identify additional folks as new topic domains are covered.  Chris and Melissa can contact me with any escalations or broader questions for resolving with the Portal team or with the Data Release groups.